Implement Level-2
Chatbot prompts user to add in tasks
and displays it to the user when the command
"list" is used.

Class Task keeps track of the
String array that stores all the tasks. The
class contains the relevant constructors, getters
and setters for the class.

Function addTask in Task.java adds the task
to the array and function printTasks prints the existing
tasks if user enters the command "list".

Initialise function in Duke .java prompts
user to enter name