Implement Level 3
The chatbot is able to update the completion statuses of the tasks that
were given by the user. The command "done " prompts the system to
change the completion status of the task as completed.

In the Task Class, the completion array stores the completion statuses of the tasks. The
completeTasks function controls the completion statuses
of the tasks.The appropriate getters and setters have been added for this class
level member

The printTasks function now accommodates the printing of
the tick and cross symbols for the task.

In the Duke Class, the commands function reads the input of the users
and decides the appropriate course of action.