Implement A-CodingStandard
Function command () causes confusion to readers.
Hence, it is changed to readInput to comply with the Coding
Standards.

Implentation of the task class as an array of objects
makes it harder for the tasks to be modified in the future. Hence, the class
now initialises single task objects. The array is then initialised
in the main function.

The functions in the Task class now modify the specific
task it is called upon to modify.