Enable application to detect the existence of data.txt file
A method has been implemented in PrintMethod.java class
to check if the data.txt file is available. The method, printContents,
will then show the tasks from a previous session.

If a file is not present, the application informs the user that a
new file will be created in the directory that contains the src folder.
The createFile() method creates the data.txt file for the user.