Implement methods to change content in data.txt file
A method, searchFile has been implemented in TaskArray.java .
It searches the existing file for the specific task and changes the task.
This is useful when the user enters the done command. Hence, the
searchFile method updates the existing file to replace the task which has
been completed.

The readTask method in TaskArray.java now reads all of the
inputs from the todo,event,deadline commands, instantiates the
task and adds the task to the array.

The createFile method in TaskArray.java creates a file if a file is
not present.

The clearFile method in TaskArray. java clears the contents of the file
if the user does not need it.

The printContents  Method has been moved from PrintMethod class
to Duke.java class.