Implement additional methods to transfer data to data.txt
The format of the  output that is being written to the
 data.txt file has been changed. The Boolean value
is now represented by 0's and 1's.  Complete tasks are assigned
with a value of 1 while incomplete tasks are assigned with a value
of 0.

The application now loads the pre existing tasks from
the data.txt file. The fileToTask method converts the
task in the data.txt file to its respective task object. After
that, the task is added in to the arraylilst.