Code Refactor for  Level 7
The directory and filepaths are now stored as constant
expressions in TaskArray.java.

The fileData method in TaskArray.java reads the contents
of data.txt and stores them as an Arraylist.

The printContents method in Duke.java class
now shows the path in which the directory would be created eventually.