Extract closely related code
The code has been refactored to extract out
the methods that perform a specific task.

The Ui Class takes care of User-interactions.
The Parser Class performs the relevant operations on
the String to extract out the task.

The storage class takes care of the loading of the data from the
data.txt file.

The taskList class modifies the ArrayList which tracks all
of the tasks.