Refactor code in AddCommand.java and Storage.java
In the AddCommand.java, the returnEventDeadline() method
returns the deadline or the event object, after instantiating it
with the appropriate time and date variable.It also finds
the date and time of the task given to it.

In the Storage.java class, the updateFile() method
is modified so that the completion status of the task
will be updated accordingly in the data.txt file.