Code refactoring to read input from File
The method taskFromFile() in storage.java now reads the date and Time input
from the files and updates it accordingly to the taskArrayList. The parseString
method in AddCommand.java has been changed to accomodate reading of file.

The classes Deadline.java and Event.java have been refactored
to implement the correct toString() method. Now, the time is shown as
HH:mm and the date is shown as dd MMM yyyy.